% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptest_old.R
\name{adaptest_old}
\alias{adaptest_old}
\title{OLD Data-Adaptive Algorithm Implementation (for reference only)}
\usage{
adaptest_old(Y, A, W = NULL, n_top, n_fold, folds_vec = NULL,
  parameter_wrapper = adaptest::rank_DE, SL_lib = c("SL.glm", "SL.step",
  "SL.glm.interaction", "SL.gam", "SL.earth"), absolute = FALSE,
  negative = FALSE)
}
\arguments{
\item{Y}{continuous or binary outcome variable}

\item{A}{binary treatment indicator: \code{1} = treatment, \code{0} = control}

\item{W}{matrix containing baseline covariates}

\item{n_top}{integer value for the number of candidate covariates to generate
using the data-adaptive estimation algorithm}

\item{n_fold}{integer number of folds to be used for cross-validation}

\item{folds_vec}{Vector of \code{numeric} indicating the validation fold in
which a given observation falls for a standard V-fold cross-validation
procedure. Default is \code{NULL}, in which case a custom cross-validation
procedure is used.}

\item{parameter_wrapper}{function}

\item{SL_lib}{character}

\item{absolute}{boolean: \code{TRUE} = test for absolute effect size. This
\code{FALSE} = test for directional effect. This overrides argument
\code{negative}.}

\item{negative}{boolean: \code{TRUE} = test for negative effect size,
\code{FALSE} = test for positive effect size}
}
\value{
\code{S3} object of class "data_adapt" for data-adaptive multiple
 testing.
}
\description{
Performs targeted minimum loss-based estimation (TMLE )of a marginal additive
treatment effect of a binary point treatment on an outcome. The data-adaptive
algorithm is used to perform variable reduction to avoid the disadvantages
associated with multiple testing. INTERNAL USE ONLY.
}
\author{
Wilson Cai \email{wcai@berkeley.edu}, in collaboration with Alan E.
        Hubbard, with contributions from Nima S. Hejazi.
}
