% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptmle.R
\docType{class}
\name{adapTMLE-class}
\alias{adapTMLE-class}
\alias{.adaptmle}
\title{Constructor for class adaptmle}
\value{
class \code{adaptmle} object, sub-classed from SummarizedExperiment.
}
\description{
Constructor for class adaptmle
}
