% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioadaptest.R
\name{bioadaptest}
\alias{bioadaptest}
\title{Data Adaptive Multiple Testing for Computational Biology}
\usage{
bioadaptest(data_in, var_int, cntrl_set = NULL, n_top = 25, n_fold = 10,
  parameter_wrapper = rank_DE, SL_lib = c("SL.glm", "SL.step",
  "SL.glm.interaction", "SL.gam", "SL.earth", "SL.mean"), absolute = FALSE,
  negative = FALSE, p_cutoff = 0.05, q_cutoff = 0.05)
}
\arguments{
\item{data_in}{...}

\item{var_int}{...}

\item{cntrl_set}{...}

\item{n_top}{...}

\item{n_fold}{...}

\item{parameter_wrapper}{...}

\item{SL_lib}{...}

\item{absolute}{...}

\item{negative}{...}

\item{p_cutoff}{...}

\item{q_cutoff}{...}
}
\value{
...
}
\description{
A thin wrapper that implements the main data-adaptive multiple hypothesis
testing strategy for data structures commonly found in computational biology
experiments, using the popular SummarizedExperiment container class.
}
