% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptest.R
\name{cv_param_est}
\alias{cv_param_est}
\title{Compute data-adaptive parameter estimate for a single cross-validation fold}
\usage{
cv_param_est(fold, data, parameter_wrapper, absolute, negative, n_top, SL_lib,
  Y_name, A_name, W_name)
}
\arguments{
\item{fold}{fold output from \code{origami}}

\item{data}{entire training data}

\item{parameter_wrapper}{user-defined function}

\item{absolute}{boolean: \code{TRUE} = test for absolute effect size. This
\code{FALSE} = test for directional effect. This overrides argument
\code{negative}.}

\item{negative}{boolean: \code{TRUE} = test for negative effect size,
\code{FALSE} = test for positive effect size}

\item{n_top}{integer value for the number of candidate covariates to generate
using the data-adaptive estimation algorithm}

\item{SL_lib}{character of \code{SuperLearner} library}

\item{Y_name}{string of \code{colnames} that all biomarkers share}

\item{A_name}{string of \code{colnames} of treatment}

\item{W_name}{string of \code{colnames} that all baeline covariates share}
}
\description{
Compute data-adaptive parameter estimate for a single cross-validation fold
}
