% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_adapt.multi_test.R
\name{data_adapt.multi_test}
\alias{data_adapt.multi_test}
\title{Data adaptive test statistic for high dimensional multiple testing}
\usage{
data_adapt.multi_test(Y, A, W = NULL, n.top.want, n.fold, abs = FALSE,
  negative = FALSE, parallel = FALSE)
}
\arguments{
\item{Y}{continuous or binary outcome variable}

\item{A}{binary treatment indicator, \code{1} - treatment, \code{0} - control}

\item{W}{vector, matrix, or dataframe containing baseline covariates}

\item{n.top.want}{integer value for the number of candidate covariates to generate from data-adaptive algorithm}

\item{n.fold}{integer number of folds for cross-validation}

\item{abs}{boolean. \code{TRUE} - test for absolute effect size. Overrides argument \code{negative}; \code{FALSE} - test for positive/negative effect size. See argument \code{negative}.}

\item{negative}{boolean. \code{TRUE} - test for negative effect size; \code{FALSE} - test for positive effect size.}

\item{parallel}{boolean. \code{TRUE} - use multi-core processing; \code{FALSE} - single core process.}
}
\description{
Carries out targeted maximum likelihood estimation of marginal additive
treatment effect of a binary point treatment on an outcome. Data-adaptive algorithm is
carried out to perform variable reduction to avoid some of the pitfalls of many multiple comparisons.
}
\examples{

}
\author{
Wilson Cai \email{wcai@berkeley.edu}, in collaboration with Alan E. Hubbard.
}

