% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_adapt_multi_test.R
\name{data_adapt_multi_test}
\alias{data_adapt_multi_test}
\title{Data-adaptive test statistics for high-dimensional multiple testing}
\usage{
data_adapt_multi_test(Y, A, W = NULL, n_top, n_fold, absolute = FALSE,
  negative = FALSE, parallel = FALSE)
}
\arguments{
\item{Y}{continuous or binary outcome variable}

\item{A}{binary treatment indicator: \code{1} = treatment, \code{0} = control}

\item{W}{matrix containing baseline covariates}

\item{n_top}{integer value for the number of candidate covariates to generate
using the data-adaptive estimation algorithm}

\item{n_fold}{integer number of folds to be used for cross-validation}

\item{absolute}{boolean: \code{TRUE} = test for absolute effect size. This
\code{FALSE} = test for directional effect. This overrides argument
\code{negative}.}

\item{negative}{boolean: \code{TRUE} = test for negative effect size,
\code{FALSE} = test for positive effect size}

\item{parallel}{boolean: \code{TRUE} = use multiple cores via \code{foreach},
\code{FALSE} = single-core processing.}
}
\description{
Performs targeted minimum loss-based estimation (TMLE )of a marginal additive
treatment effect of a binary point treatment on an outcome. The data-adaptive
algorithm is used to perform variable reduction to avoid the disadvantages
associated with multiple testing.
}
\author{
Wilson Cai \email{wcai@berkeley.edu}, in collaboration with Alan E.
        Hubbard, with contributions from Nima S. Hejazi.
}
