% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_composition.R
\name{get_composition}
\alias{get_composition}
\title{Generic method of \code{table} for objects of class \code{data_adapt}}
\usage{
get_composition(adapt.fit, type = "small")
}
\arguments{
\item{adapt.fit}{data-adaptive statistical object of class \code{data_adapt}
as returned by \code{data_adapt_multi_test}}
}
\description{
Customized informative tables for examining data-adaptive statistics
}
