% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptest.R
\name{get_pval}
\alias{get_pval}
\title{Statistical Inference for Data-Adaptive Parameters}
\usage{
get_pval(Psi_output, EIC_est_final, alpha = 0.05)
}
\arguments{
\item{Psi_output}{vector containing differential expression estimates}

\item{EIC_est_final}{matrix where each column is the efficient influence curves of the \code{Psi_output}}

\item{alpha}{floating number between (0,1), significance level of the test}
}
\value{
\code{pval} p-values for each \code{Psi_output}

\code{upper} upper confidence bound for each \code{Psi_output}

\code{lower} lower confidence bound for each \code{Psi_output}

\code{sd_by_col} standard error of efficient influence curve for each \code{Psi_output}
}
\description{
Statistical Inference for Data-Adaptive Parameters
}
