% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptmle-class.R
\name{get_results_adaptmle}
\alias{get_results_adaptmle}
\title{Accessor for results of class adaptest}
\usage{
get_results_adaptmle(adaptmle_in, data_adapt_out)
}
\arguments{
\item{adaptmle_in}{S4 object of class{adapTMLE} into which analytic results
should be inserted for organizational and interfacing purposes.}

\item{data_adapt_out}{S3 object of class \code{data_adapt} with analytic
results from the statistical routine, as produced by \code{adaptest}.}
}
\value{
Called for side-effects (i.e., altering \code{adaptmle} objects).
}
\description{
Simple accessor function to extract and assign important statistical results
 from the main function \code{adaptest}. INTERNAL USE ONLY.
}
\keyword{internal}
