% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_parameter_wrapper.R
\name{rank_DE}
\alias{rank_DE}
\title{Ranking for data-adaptive test statistics}
\usage{
rank_DE(Y, A, W, absolute = FALSE, negative = FALSE)
}
\arguments{
\item{Y}{continuous or binary outcome variable}

\item{A}{binary treatment indicator: \code{1} = treatment, \code{0} = control}

\item{W}{vector, matrix, or data.frame containing baseline covariates}

\item{absolute}{boolean: \code{TRUE} = test for absolute effect size. This
\code{FALSE} = test for directional effect. This overrides argument
\code{negative}.}

\item{negative}{boolean: \code{TRUE} = test for negative effect size,
\code{FALSE} = test for positive effect size}
}
\description{
Performs ranking using Targeted Minimum Loss-Based Estimation. This function
is designed to be called inside \code{adaptest}; it should not be run by
itself outside of that contex.
}
