% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_parameter_wrapper.R
\name{rank_ttest}
\alias{rank_ttest}
\title{Compute ranking of biomarkers by sorting t-test p-values}
\usage{
rank_ttest(Y, A, W)
}
\arguments{
\item{Y}{(numeric vector) - continuous or binary biomarkers outcome variables}

\item{A}{(numeric vector) - binary treatment indicator: \code{1} = treatment,
\code{0} = control}

\item{W}{(numeric vector, numeric matrix, or numeric data.frame) - matrix of
baseline covariates where each column corrspond to one baseline covariate
and each row correspond to one observation.}
}
\value{
an \code{integer vector} containing ranks of biomarkers.
}
\description{
Compute ranking of biomarkers by sorting t-test p-values
}
\examples{
set.seed(1234)
data(simpleArray)
rank_ttest(Y = simulated_array,
           A = simulated_treatment,
           W = rep(1, length(A)))
}
