% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_treatment}
\alias{simulated_treatment}
\title{Simulated differential expression data with one exposure}
\format{A numeric vector containing binary exposures
\describe{
  This is example data to be used in testing the \code{adaptest}
  procedure. Consult the vignettes for how to use this data.
}}
\usage{
simulated_treatment
}
\value{
A numeric vector \code{simulated_treatment}.
}
\description{
A dataset containing 1e4 biomarkers and one exposure
}
\keyword{datasets}
