% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{[<-.result}
\alias{[<-.result}
\title{multiple assign
Citation: ggrothendieck/gsubfn/}
\usage{
\method{[}{result}(x, ...) <- value
}
\arguments{
\item{value}{}
}
\description{
multiple assign
Citation: ggrothendieck/gsubfn/
}
\examples{
#NA
}
