% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.data_adapt.R
\name{table.data_adapt}
\alias{table.data_adapt}
\title{Generic method of \code{table} for objects of class \code{data_adapt}}
\usage{
table.data_adapt(adapt.fit)
}
\arguments{
\item{adapt.fit}{data-adaptive statistical object of class \code{data_adapt}
as returned by \code{data_adapt_multi_test}}
}
\description{
Customized informative tables for examining data-adaptive statistics
}
