% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Y}
\alias{Y}
\title{Simulated differential expression data with one exposure}
\format{A numeric matrix containing 1e4 biomarkers of 1e2 subjects.
\describe{
  This is example data to be used in testing the \code{adaptest}
  procedure. Consult the vignettes for how to use this data.
}}
\usage{
Y
}
\value{
A matrix \code{Y}
}
\description{
A dataset containing 1e4 biomarkers and one exposure
}
\keyword{datasets}
