% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptest.R
\name{adaptest}
\alias{adaptest}
\title{Data-adaptive Statistics for High-Dimensional Multiple Testing}
\usage{
adaptest(Y, A, W = NULL, n_top, n_fold,
  parameter_wrapper = adaptest::rank_DE, SL_lib = c("SL.glm", "SL.step",
  "SL.glm.interaction", "SL.gam", "SL.earth"), absolute = FALSE,
  negative = FALSE, p_cutoff = 0.05, q_cutoff = 0.05)
}
\arguments{
\item{Y}{continuous or binary outcome variable}

\item{A}{binary treatment indicator: \code{1} = treatment, \code{0} = control}

\item{W}{matrix containing baseline covariates}

\item{n_top}{integer value for the number of candidate covariates to generate
using the data-adaptive estimation algorithm}

\item{n_fold}{integer number of folds to be used for cross-validation}

\item{parameter_wrapper}{function}

\item{SL_lib}{character}

\item{absolute}{boolean: \code{TRUE} = test for absolute effect size. This
\code{FALSE} = test for directional effect. This overrides argument
\code{negative}.}

\item{negative}{boolean: \code{TRUE} = test for negative effect size,
\code{FALSE} = test for positive effect size}

\item{p_cutoff}{A \code{numeric} indicating the cutoff at and below which
p-values are to be considered significant.}

\item{q_cutoff}{A \code{numeric} indicating the cutoff at and below which
q-values (p-values after multiple testing corrections are applied) are to be
considered significant.}
}
\description{
Performs targeted minimum loss-based estimation (TMLE )of a marginal additive
treatment effect of a binary point treatment on an outcome. The data-adaptive
algorithm is used to perform variable reduction to avoid the disadvantages
associated with multiple testing.
}
