% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_adapt.R
\name{data_adapt}
\alias{data_adapt}
\title{Constructor function for class \code{data_adapt}}
\usage{
data_adapt(Y, A, W = NULL, n.top, n.fold, absolute, negative, parallel)
}
\arguments{
\item{Y}{continuous or binary outcome variable}

\item{A}{binary treatment indicator: \code{1} = treatment, \code{0} = control}

\item{W}{vector, matrix, or data.frame containing baseline covariates}

\item{n.top}{integer value for the number of candidate covariates to generate
using the data-adaptive estimation algorithm}

\item{n.fold}{integer number of folds to be used for cross-validation}

\item{absolute}{boolean: \code{TRUE} = test for absolute effect size. This
\code{FALSE} = test for directional effect. This overrides argument
\code{negative}.}

\item{negative}{boolean: \code{TRUE} = test for negative effect size,
\code{FALSE} = test for positive effect size}

\item{parallel}{boolean: \code{TRUE} = use multiple cores via \code{foreach},
\code{FALSE} = single-core processing.}
}
\value{
S3 object of class "data_adapt" for data-adaptive multiple testing.
}
\description{
Constructor function for class \code{data_adapt}
}

