% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptest.R
\name{data_adapt}
\alias{data_adapt}
\title{Class Constructor for Data Adaptive Parameters}
\usage{
data_adapt(Y, A, W = NULL, n_top, n_fold, absolute, negative,
  parameter_wrapper, SL_lib)
}
\arguments{
\item{Y}{continuous or binary outcome variable}

\item{A}{binary treatment indicator: \code{1} = treatment, \code{0} = control}

\item{W}{vector, matrix, or data.frame containing baseline covariates}

\item{n_top}{integer value for the number of candidate covariates to generate
using the data-adaptive estimation algorithm}

\item{n_fold}{integer number of folds to be used for cross-validation}

\item{absolute}{boolean: \code{TRUE} = test for absolute effect size. This
\code{FALSE} = test for directional effect. This overrides argument
\code{negative}.}

\item{negative}{boolean: \code{TRUE} = test for negative effect size,
\code{FALSE} = test for positive effect size}

\item{parameter_wrapper}{user-defined function}

\item{SL_lib}{character of SuperLearner library}
}
\value{
\code{S3} object of class "data_adapt" for data-adaptive multiple testing.
}
\description{
Class Constructor for Data Adaptive Parameters
}
