% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary.R
\name{get_composition}
\alias{get_composition}
\title{Generic method of \code{table} for objects of class \code{data_adapt}}
\usage{
get_composition(object, type = "small")
}
\arguments{
\item{object}{data-adaptive statistical object of class \code{data_adapt}
as returned by \code{adaptest}}

\item{type}{string. 'small' or 'big'. 'small' mode returns composition of significant data-adaptive parameters. 'big' mode returns composition of all data-adaptive parameters}
}
\description{
Customized informative tables for examining data-adaptive statistics
}
