% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptest.R
\name{get_pval}
\alias{get_pval}
\title{Statistical Inference for Data-Adaptive Parameters}
\usage{
get_pval(Psi_output, EIC_est_final, alpha = 0.05)
}
\arguments{
\item{Psi_output}{...}

\item{EIC_est_final}{...}

\item{alpha}{...}
}
\description{
Statistical Inference for Data-Adaptive Parameters
}
