% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary.R
\name{get_significant_biomarker}
\alias{get_significant_biomarker}
\title{Extract statistically significant biomarkers}
\usage{
get_significant_biomarker(object, cutoff = 0.5)
}
\arguments{
\item{object}{\code{data_adapt} object}

\item{cutoff}{cut-off value for composition percentage}
}
\value{
(integer vector) of significant gene index
}
\description{
Extract statistically significant biomarkers
}
\examples{
set.seed(1234)
data(simpleArray)
simulated_array <- simulated_array
simulated_treatment <- simulated_treatment

adaptest_out <- adaptest(Y = simulated_array,
                         A = simulated_treatment,
                         W = NULL,
                         n_top = 5,
                         n_fold = 3,
                         learning_library = 'SL.glm',
                         parameter_wrapper = adaptest::rank_DE,
                         absolute = FALSE,
                         negative = FALSE)
get_significant_biomarker(adaptest_out)
}
