% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary.R
\name{get_significant_biomarker}
\alias{get_significant_biomarker}
\title{Extract statistically significant biomarkers}
\usage{
get_significant_biomarker(object)
}
\arguments{
\item{object}{\code{data_adapt} object}
}
\description{
Extract statistically significant biomarkers
}
\keyword{internal}
