% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.data_adapt.R
\name{plot.data_adapt}
\alias{plot.data_adapt}
\title{#' Generic method of \code{plot} for objects of class \code{data_adapt}}
\usage{
\method{plot}{data_adapt}(adapt.fit, plot_id = c(1, 2, 3))
}
\description{
Customized plotting method for easily examining data-adaptive statistics
}

