% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_plot_shinyprint_table.R
\name{plot.data_adapt}
\alias{plot.data_adapt}
\title{Generic plot method for data_adapt class}
\usage{
\method{plot}{data_adapt}(x, ..., plot_type = c("cvrank", "pvals"))
}
\arguments{
\item{x}{data-adaptive statistical object of class \code{data_adapt} as
returned by \code{adaptest}}

\item{...}{additional arguments passed to \code{plot} as necessary}

\item{plot_type}{character vector specifying which of the two types of plots
to generate: "cvrank" for a plot sorted average CV-rank, or "pvals"
for a plot sorted by p-values with labels corresponding to indices}
}
\description{
Customized plotting method for easily examining data-adaptive statistics
}
