% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.data_adapt}
\alias{print.data_adapt}
\title{Generic print method for data_adapt class}
\usage{
\method{print}{data_adapt}(x, ...)
}
\arguments{
\item{x}{(data_adapt) - object of class \code{data_adapt} as returned by
\code{adaptest}}

\item{...}{additional arguments passed to \code{print} as necessary}
}
\value{
strings into stdout; containing information of the fitted model
}
\description{
Customized informative print method for examining data-adaptive statistics
}
