% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.data_adapt.R
\name{print.data_adapt}
\alias{print.data_adapt}
\title{Generic method of \code{print} for objects of class \code{data_adapt}}
\usage{
\method{print}{data_adapt}(adapt.fit, ...)
}
\arguments{
\item{adapt.fit}{data-adaptive statistical object of class \code{data_adapt}
as returned by \code{data_adapt_multi_test}}
}
\description{
Customized informative print method for examining data-adaptive statistics
}

