% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyprint.data_adapt.R
\name{shinyprint.data_adapt}
\alias{shinyprint.data_adapt}
\title{Method of \code{shinyprint} for objects of class \code{data_adapt}}
\usage{
shinyprint.data_adapt(adapt.fit)
}
\arguments{
\item{adapt.fit}{data-adaptive statistical object of class \code{data_adapt}
as returned by \code{data_adapt_multi_test}}
}
\description{
Provides HTML-based printing utility for examining data-adaptive statistics
}

