% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shinyprint.data_adapt}
\alias{shinyprint.data_adapt}
\title{Method of \code{shinyprint} for objects of class \code{data_adapt}}
\usage{
shinyprint.data_adapt(x)
}
\arguments{
\item{x}{data-adaptive statistical object of class \code{data_adapt}
as returned by \code{adaptest}}
}
\description{
Provides HTML-based printing utility for examining data-adaptive statistics
}
