% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.data_adapt}
\alias{summary.data_adapt}
\title{Summary tables of the `adaptest` object}
\usage{
\method{summary}{data_adapt}(object, ..., type = "adapt_param")
}
\arguments{
\item{object}{(data_adapt) object as returned by \code{adaptest}}

\item{...}{not implemented}

\item{type}{(character) - `adapt_param` or `biomarker`.
`adapt_param` mode summarizes the data-adaptive target
parameter. `biomarker` mode summarizes chracteristics of the biomarkers from
the original data}
}
\value{
(data.frame) of the summary statistics

\code{type = 'adapt_param'} wtih columns: 'data-adaptive parameters', 'Differential expression', 'p-values', 'q-values'

\code{type = 'biomarker'} wtih columns: 'biomakers', 'mean rank', '% appear in top'
}
\description{
Summary tables of the `adaptest` object
}
